requirements.txt
cryptography==36.0.2
Pillow==9.1.0

scrypt~=0.8.20
numpy~=1.22.3
pandas~=1.4.2
opencv-python~=4.5.5.64